refactor(filters): extract shared filter_conditions module
Move the operator list, per-fieldtype invalid-condition map, value-control
mapping (set_fieldtype), timespan options, and value helpers out of the jQuery
filter UI into a shared ES module, so the legacy filter and the new advanced
filter UI share one source of truth and cannot drift.

- filter_conditions.js: new shared module, also exposed on
  frappe.ui.filter_conditions / frappe.ui.filter_utils. Besides the moved
  definitions it adds get_valid_conditions_for_field(), get_fieldtype_override(),
  is_nested_set_field(), and prepare_filter_docfield() for reuse by both UIs.
- filter.js: consume the shared module throughout - set_conditions,
  set_conditions_from_config, the operator-change fieldtype override, the
  nested-set toggle, and the filter docfield prep all delegate to it, leaving no
  inline duplicate of the extracted logic. Behaviour is unchanged.
